slack-bot/requirements.txt
slackclient
Flask==0.12.2
paho-mqtt==1.4.0
google-cloud-tasks==0.3.0
google-cloud-datastore